+++ backend/app/models/restaurant.rb
@@ -14,8 +14,6 @@ class Restaurant < ApplicationRecord
              primary_key: :id
 
   def self.in_bounds(bounds)
-    puts '*************'
-    puts bounds
     where('latitude < ?', bounds[:northEast][:latitude])
       .where('latitude > ?', bounds[:sourthWest][:latitude])
       .where('longitude < ?', bounds[:northEast][:longitude])
